eeg_speech_pipeline/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
PyWavelets>=1.1.0
faiss-cpu>=1.7.0
openai>=1.0.0
pyyaml>=6.0
